checks: fix three ruff findings in the new oracle scripts
Caught by ci.yml's Script Lint job, which I had not run locally before
pushing - ruff.toml's curated set is wider than the E/W/F/I I assumed, and
includes RUF and PL rules:

  derive_channel_floors.py  RUF046  int() around math.floor(), which already
                                    returns an int in Python 3
  test_compare_wav.py       RUF100  an unused `# noqa: E402` - E402 is not in
                                    the selected set, so the suppression was
                                    suppressing nothing
  test_compare_wav.py       PLW2901 rebinding the `with` target (`tmp`) to the
                                    Path built from it

None changes behaviour; the 12 tests pass unchanged. Verified this time with
CI's own pinned linters (ruff 0.16.4 from requirements/requirements-lint.txt,
installed --require-hashes) rather than by inspection: `ruff check .` is clean
over the whole tree, shellcheck is clean over every tracked .sh, and actionlint
accepts the workflow change.

+++ tools/checks/derive_channel_floors.py
@@ -94,7 +94,7 @@ def derive(records: list[dict]) -> tuple[list[float], list[int], int]:
     channels = len(records[-1]["channels_db"])
     usable = [r for r in records if len(r["channels_db"]) == channels]
     minima = [min(r["channels_db"][i] for r in usable) for i in range(channels)]
-    return minima, [int(math.floor(m - HEADROOM_DB)) for m in minima], channels
+    return minima, [math.floor(m - HEADROOM_DB) for m in minima], channels
 
 
 def main() -> int:
